requirements.txt
# Core LLM & AI
groq
langchain
langchain-groq
langchain-community

# OpenAI
openai

# Anthropic
anthropic

# Google Gemini
google-generativeai
langchain-google-genai

# Hugging Face
transformers
huggingface-hub
tokenizers
datasets
accelerate

# Data Processing
pandas
openpyxl
numpy

# Machine Learning Utilities
scikit-learn
scipy
joblib

# Visualization
matplotlib
plotly

# Web UI
streamlit
flask
flask-cors

# Environment Variable Management
python-dotenv
python-decouple
pydantic-settings

google-auth
requests
razorpay
gunicorn
